backend/requirements.txt
Flask
PyJWT
flask-cors
boto3
